Allow PNF service chaining to work with RTF
Route re-origination for PNFs may not work with RTF because CN
does not advertise RTarget routes for RTs for which there's no
subscriptions from xmpp agents. As a result, service chaining
use cases with PNFs where the the source and destination of the
traffic are outside the cluster/DC are impacted.

Fix by introducing flag that makes CN advertise RTarget routes
for an instance unconditionally i.e. even if there are no xmpp
agents subscribed to the instance.  The flag will be updated by
ST based on whether there are any policies with PNF services
applied to the VN.

Implementation Details

- Add has_pnf bool to BgpInstanceConfig and update it from config
- Add always_subscribe bool to RoutingInstance to track has_pnf
- Add and delete Local RTarget routes for import RTs of instance
- Use instance index as path id for Local RTarget routes
- Handle changes in always_subscribe bool for an instance
- Handle changes in import RTs for an instance
- Handle changes in ASN and RouterId

Change-Id: Ie9d8c9a652eae2682be21226062d6ffb269c308d
Closes-Bug: 1496695