Use vm_name to delete the VM object
Schema transformer was incorrecty using si_name to delete VM object. It should
use vm_name. Because of this, stale VM objects remain in schema transformer
local memory and so, if SI is created again, it thinks the VM exists but no VMIs
exist and hence service chain is not created.

This is partial fix because we still do not break service chain when SI is deleted.

Change-Id: I4458352e3af0feeb1f68c57920b6875991cfac81
Partial-Bug: 1499125