* If a ECMP flow gets processed as layer-2 flow mark it as short flow
If source route is composite, then the flow has to be layer3
procesess so that proper load distribution happens, this is
forced but proxying with vrouter mac for ECMP destinations
and if ECMP source originates the packet then also vrouter
would proxy with vrrp mac so that packet gets processed
as layer3 flow. If a flow transition from non-ECMP to
ECMP then for some time flow would be layer2. Mark such flow
as short flow for that time.
Closes-bug:#1472796

Change-Id: If2195c8add7d31f32907026e7e53af56614f1693