Destroy service chain if service instance is deleted
Currently, once a service chain is created, we do not check again until the service instance is removed from policy or policy is detached from VN. But it is possible that either the VM, VMI or service instance has been deleted or has some problem. In those cases, we should check the validity of the service chain again and destroy it if everything is no longer valid.

Change-Id: Ic16d4af50057f6e842d1d4e2356d72728681a182
Closes-Bug: 1499125