Fix end of rib logic in channel to handle all message types.
Currently, to determine that we have received all data, we were depending the
message type 'SearchResult'.

But, there can be cases, where the first batch of data is received in an
'UpdateResult' i.e. when the connection to the ifmap-server goes up,
ifmap-server does not have any data to begin with. Then, later when config is
added, the data will come in as an 'UpdateResult'. The code is processing the
EOR logic only for SearchResults which is incorrect. Fixing that.

Change-Id: I0106c62e8bec2cbf06a25efc1856451273d7df1f
Closes-Bug: #1501425