Set MED for active/backup routes sent to eBGP peers
Existing implementation sends local preference to iBGP peers but does
not work for eBGP. Fix is to automatically calculate the med based on
local preference sent by vrouter agent. This is done only if agent has
not explicitly sent med to avoid problems in case agent needs to send
med explicitly in future.

Change-Id: I484e1d6bbf32cd1f21d572c5666201eac8ca3bc3
Closes-Bug: 1501412