Add has_pnf property to routing instances as required
When a pnf service instance is added to a service chain, the corresponding service routing instance and
primary routing instance need to have has_pnf property set. Similarly, the property needs to be unset
when no pnf service instances are attached to the network.

Also added test case for the same but it is disabled for now because it needs service monitor changes
which are still not committed.

Change-Id: I806345df6e57b4e497282dfe490c4bb344925ddf
Closes-Bug: 1496695