Enable Tcp keepalive to detect peer DOWN on XmppServer side.
Configured tcp_hold_time = x, we will set the keepalive as x/2 for
x>18secs. Mininum value of x defaults to 9secs.
From experiments it is seen the detection takes twice the time in case
there is a pending data (here XmppKeepalive) in the tcp socket.

Disable support on XmppClient end as the vrouter can take longer to
detect the Xmpp Server is down and hence will retain routes atleast
for 90secs.

Change-Id: Ie20150df6c76b355233a56779a5e4811a2fa6f37
Closes-Bug:1479198