Enable Tcp keepalive to detect peer DOWN on XmppServer side.
Configured tcp_hold_time = x, we will set the keepalive as x/2 for x>25secs.
From experiments it is seen the detection takes twice the time in case
there is a pending data (here XmppKeepalive) in the tcp socket. Hence
the range can be between 15+/- to 30+/-, but median detection is always 30secs.
Xmpp Keepalive is every 30sec and hence the number.

Change-Id: Ie20150df6c76b355233a56779a5e4811a2fa6f37
Closes-Bug:1479198