Comparing service instance config properties in OnChange
Right now when there is properties change in service instance config
object, the properties are calcualted  in IFNodeToReq and these parameters are
compared with operdb parameters. Only if these parameters are different
they are enqueued to OperDB. If not the change is not enqueued.

The above thing can result in a bug in the following scenario.
1) svc_instance has config A
2) it is modified to config B. As there is a change B gets enqueued to
Oper. But OnChange() is not invoked yet.
3) config again changes to A. As B is not yet written to oper, when
latest config A is compared with oper, which is also A, we dont enqueue
the change any more to Oper.
4) Now OnChange() gets invoked for B and oper gets modified to B.
5) But the latest config change A never gets written to oper, leading to
bugs.

To fix the above scenario, the properties should be compared only in
OnChange() but not in IFNodeToReq.

Change-Id: Ibfb36bdcd75bc0108a9d09c6add8a4fad827d5e0
closes-bug: #1502190