Snat health check generates unnecessary api server updates
On vrouter down detach infro is not sent correctly to the api server. This causes a bunch of
messages/updates to the API server on a regular interval. The fix now correctly updates the
virtual-machine and vrouter references so that a new vrouter can be attached to the vm.

In addition made the service check interval configurable in contrail-svc-monitor.conf
check_service_interval=<seconds>

Change-Id: I352ec821128e801b24cd227ce35824906140fcab
Closes-Bug: #1502300