Floating-IP not set for LB VMs
Intermittently floating ip is not carried over from VIP
to the netns virtual machine interfaces. This is caused
when floating ip is configured on the VIP and the db messages
come out of order for service instance and floating ip.

Add code to handle fip to vmi association during launch of
service netns vm in addition to the specifiec fip handling
that already exists.

Change-Id: I73dbf1775e46f0f85fcdb293784d42e75384905c
Closes-Bug: #1501020