Detect unit-test failure in svc-monitor
All other usage of testr/run_test.sh infra does an
explicit search for \nOK as a check for unit-test
failure (in setup.py's RunTestsCommand.run()).

A couple of setup.py's were missed, and the result
of os.system() is not being checked. This results in
unit-test failures being silenlty missed.

Change-Id: I08c2bab5e6266c91f599bd1efbea44f89d655a2e
Closes-Bug: #1507070