Enable IPv6 for all tests in svc_static_route_intergration_test
The tests were previously disabled for IPv6 when mx_push_connected_
was true.

The IPv6 vpn code didn't generate a fallback RD when there's no source RD
in the vrf route. Hence the RDs used on mx when replicating the connected
routes in blue and emcp VRF tables to the v6 vpn table were 0:0. So only
one connected route got sent to cn1 and cn2.

Change-Id: Iaa2e8d3cb022b610eea308c7896c334873994f11
Partial-Bug: 1430089