Relax strict adherence to RTC RFC for eBGP peers
Current implementation of RTC (RFC 4684) advertises VPN routes to
exactly 1 eBGP peer in a given ASN even when multiple peers have
advertised an RTC route for a route target.  Expectation is that
the peer in question will advertise the VPN routes to it's iBGP
peers.  While this is technically sufficient, it could result in
sub-optimal multi-path within the neighbor AS in some scenarios.

Fix is to change implementation to send VPN routes to all peers by
default and provide a knob to limit the number of external paths in
future if/when required. IOW, the default value of the TBD knob is
infinity.

Tweak existing tests based on the above behavior change. Note that
test PathSelection3 was incorrect to begin with - there's nothing
in the RFC which says that VPN routes should only be sent to eBGP
peers and not to iBGP peers.  The behavior was just a side effect
of how the eBGP logic was implemented.

Note that JUNOS implementation of RTC advertises VPN routes to only
the best eBGP peer by default, but provides an external-paths knob
to modify this behavior.

Change-Id: I7a4101f4a63de1febaf2b56d0daa325f9042b244
Closes-Bug: 1509945