Seperate analytics database nodemgr check
If the analytics database is not present on the node, nodemgr will
not report initializing status when the cassandra data directory
for analytics is not present.

This commit adds a config option to allow specifying which
databases are present on the node, and if the analytics database
is not configured, it will not enter initializing state due to
the analytics keyspace directory's absence.

Closes-Bug:1510736

Change-Id: I587fe85ed3bd8496f57fdaab834d215d7c2d3d67