Add more functionality to the ifmap stress test
1. Add code to verify the existence of the nodes/links in the tables at the end
of the test using the data-structures maintained by IFMapStressTest.
2. There can be cases where all the events have been executed and the test is
done although the events have not been processed by the code yet. In such
cases, if the test cleans up the tables, the processing code will assert or
crash since its not finished yet. Add a config knob to wait for cpu to idle out
and let the test use that to wait before cleaning up.

Change-Id: I957248f98b0c4032843cfc95eb65f9ee3b1d7873
Partial-Bug: #1469909