Scaled service delete issue with static routes
Currently there is an issue when static routes are configured on an interface
of a service instance. The deletion is not handled properly when a service
instance is deleted. With scaling there are multiple service instance interfaces
referring to the same static route table when scaling is enabled for the
service instance. The delete of a neutron port returns an error because the
interface route table still has references to other ports of the scaled
instances. Hence heat stack delete for scaled service instances with static
routing showed stale VMs in error state.

The fix is to ignore the reference error since the last neutron port cleanup
will remove the static route table. Hence neutron port deletes should return
no error even if static route table is not deleted for shared instances.

Change-Id: Iaa49aaefe297f027ae0c78da2ac9365ed5cf9037
Closes-Bug: #1512556