Run KSyncSock from a thread instead of work-queue
Implmentation of a shared transmit queue between agent and KSync.

KSync i/o operations are done on KSync Netlink socket. Even if the socket is
set to non-blocking mode, the KSync socket i/o call will block till VRouter
completes message processing. So, netlink i/o is not done in agent context.
All messages are enqueued to a transmit and messages and i/o is done thru
the transmit queue. This ensures agent doesnt block due to ksync processing

There are two implementations of transmit queue

WorkQueue based
---------------
WorkQueue based implementation is used only for UT cases. WorkQueue based
implementation is not used in production enviroment due to performance
reasons. When we have slow producer and fast consumer, WorkQueue will
result in spawning of a "task" for every message. The task spawning happens
in producer context. In case of flow processing, agent is slow producer and
vrouter is fast consumer. As a result, using WorkQueue will result in many
task spawns thereby introducing latencies.

We retain WorkQueue based implementations since UT code depends on
WaitForIdle() APIs to continue with validations. Using Event-FD based
implementations cannot rely on WaitForIdle()

Event-FD based
--------------
This implementation is based on event_fd. The producer will add the entry
into a tbb::concurrent_queue and notify the consumer on an event_fd. The
consumer will block with "read" on event_fd. On getting an event, it will
drain the queue of all messages in a tight loop. Consumer blocks on "read"
when there is no data in the queue. This is an efficient implementation of
queue between agent and ksync

Change-Id: I50d6cf6ac16f75147c2d4b85d9c20485e8cd4330
Partial-Bug: #1499965