Dont log error in case of eof error for async_read
async_read returns eof to indicate that the other side has closed the socket.
This is a way for other side to know that socket should be closed now. There
is no need to log since it is not an error condition

Closes-Bug: 1506795

Change-Id: I5dc5dc0cb9a909d97dacff8ceef89c863a7536c7
(cherry picked from commit ef96e3dd4710387d8aed63c948a8b74875900836)