Parsing the yaml file for finding analytics dir
At present, we look for data_file_directories text in cassandra.yaml file by
using grep. This restricts yaml file to be in a particular format. As part of
this checkin, parsing of yaml file is done so that there is no restriction on the
way yaml file is written

Change-Id: I7f0ace40fe658e2d7bcac15d743b985cbb1311fe
Closes-Bug: 1479698