Schema transformer should restore route target list at restart
When schema transformer restarts, it resets route targets to what is currently
applicable. But it is not restoring the primary routing-instance's target as an
export target on service routing instances if needed. With this patch, that is
fixed.

Change-Id: Id829af2e3d16f98b52f2b21a8c9f2fdff21ecde5
Closes-Bug: 1518415
(cherry picked from commit f50ca65696205792e83f909b5eee65da00e96187)