Add option for import_export policy in add_router.py
The script now supports the --route_target_import_export option that can
be used to set the import_export policy of the route-target within the
routing-instance. Valid values are 'import', 'export' or 'None'. The
default is 'None'. 'None' correspond to import AND export.

It also can be used to modify an existing relation.

Change-Id: I345656862cca325b41cee834d76b518cd38f057f
Closes-Bug: #1519889