Allow unidirectional connections between routing instances
Add a destination-instance field to ConnectionType to specify the
direction of a connection.  An empty value is used to indicate a
bidirectional connection. If there's a connection betwen RI-A and
RI-B and the destination-instance is set to RI-A, then routes are
imported from RI-B into RI-A, but not in the other direction.

Highlights:

- Schema change to make connection a link with properties
- Modify ifmap config code to take above schema changes into account
- Update dependency tracker policy in bgp_config_listener.cc
- Update and enhance tests in bgp_config_listener_test.cc
- Enhance tests in bgp_config_test.cc

Change-Id: Ic4d40df47af57b8bf990a7d3020ec7f70dc858d0
Closes-Bug: 1518033