Advertise origin vn for each ecmp path in xmpp
This capability is required for advanced service chaining use cases
where the same route is learnt from 2 or more service chains. An ACL
tells the vRouter which chain to use for given flow.

Temporarily disable xmpp_ecmp_test. Ananth is working on updating it
to properly handle multiple origin vns for a path and not resuse the
same label for in different VRFs.

Highlights:

- Schema channge to add virtual-network to NextHopType
- Move some code from bgp_ribout.h to bgp_ribout.cc
- Get rid of vrf_originated_ field in RibOutAttr
- Encode the origin_vn_index in each NextHop
- An origin_vn_index of 0 says that path is vrf originated
- XmppMessageBuilder translates origin_vn_index to a name
- Sprinkle const in a few places
- Add integration tests to verify new functionality

Change-Id: I72c8aab47ba1ab9141a811a89a93ed22114802e4
Closes-Bug: 1518034