Send Evicted flow stats from agent.
Vrouter informs agent about stats of evicted from when the index is reused by some other flow.
The stats are sent along with flow trap message and index update message to agent. If the flow is
already deleted from agent when the stats message comes, we don't update the stats.

Change-Id: Ibee7efb0c302f6cd32e30c38d5e50f3786f7a137
Partial-Bug: #1521872