Port delete handling for shared instance-ip refs
For port tuple a new reference has been added from service instance
to instance-ip. Hence delete of port fails because of the back-refs
from service instance in case of shared-ip.

Drop the last vmi reference if the delete fails. The creator of the
shared instance-ip is responsible for deleting this instance-ip.

Change-Id: If746a861e2ab225ecbd57926fbdccf8977a76ef4
Partial-Bug: #1512837