Specify the full path of the sandesh packages in init_generator()
In the call to init_generator(), sandesh packages list is passed as one
of the parameneters. Presently, the relative path of the package is
being passed ['sandesh']. Therefore, the sandesh library is unable to
import the package and logs error.

Passing the full path of the sandesh package ['discovery.sandesh'] in
the packages list to fix the import error.

Change-Id: I05d120b58a5b8dcf445463d9e1f86a1035702b79
Related-bug: #1524517