Agent crashed because of freed xmpp channel.
Problem:
When agent xmpp channel is deleted, event to bring down bgp peer associated with
same is issued. This in turn starts a walk to clear all states and path
associated with this bgp peer. On finishing walk it unregisters all table
listeners. Meanwhile there can be a notification to controller on deleted peer.
This notification uses channel to verify if peer is deleted and ignore
notification. In this case channel was also deleted and hence crash happens.

Solution:
Defer deletion of channel till bgp peer is deleted.
This will not be executed if channel is not deleted but keeps flapping.

Change-Id: Ia43cbda1d9d2e8f30f30f36165b967230ec39025