Remove unnecessary socket calls and update key cache during config processing.
Since we recycle the bgp session when the authentication key is changed, we can
remove socket calls on the session socket when the key is changed since that
socket will be closed shortly after the key configuration is processed.
Instead, we will set the key on the session socket when either the session is
newly created or a new connection is accepted.
Also, only the configuration processing code should save the chosen key in the
cache (inuse_auth_key_). The code creating the new session or accepting a new
session must just used the cached value. Remove the calls to InstallAuthKeys()
from that code.

Change-Id: Ic57d9d8d7825cce5ab95a06be6c12985c2e0539d
Closes-Bug: #1526503