Stale label in TSN.
Problem:
In RecomputeMulticastPath evpn label was populated when
local_vm_peer path was not present. This was happening as it was populated from
other paths (tor/evpn) even in absence of local_vm_peer_path. This results in
stale label.
Example: No local vm peer path is present, tor path is present for flood route.
Now fabric path was added. RecomputeMulticastPath will populate evpn_label due
to presence of tor path. Addition of label has a check to add only if evpn label
is not zero and path peer is not BGP. In this case label was populated from tor
and peer is fabric_peer. Hence a stale label is added.

Solution:
Add label only when local vm peer path is added.

Change-Id: Ibc7c89b890ee7ee3a9ccc163b9cce086d278e98f
Closes-bug: 1525737