disc_server.py: remove hardcoded IP-address
src/discovery/disc_server.py has a hardcoded IP address
(127.0.0.1) in its "connect-to-self" code.

The Discovery Server can bind to a configurable port number, so this is
simply a forgotten variable.  Without this change, it is impossible to
operate a Discovery Server that has bound to another IP address than
127.0.0.1.

The default value in self._args.listen_ip_addr is populated by the
parameter "_WEB_HOST" in src/discovery/disc_consts.py:5.
The default value of this parameter is '127.0.0.1'.

The change has therefore no impact on a Discovery Server with standard
configuration.

Change-Id: Ibd6a7ddd31a462f240c2358d657e5d384a5bd3ed
Closes-Bug: #1531214
Signed-off-by: Martin Millnert <martin@millnert.se>