When multiple config nodes are present, the ContrailConfig UVE struct can be a list-of-list instead of a dict.
This is because we are currently unable to detect if all structs
are equal. We would have to use OrderedDict and do a bunch of
intermediate sorting to do this.

To get around this, the alarm evaluation code will only one
of the values of ContrailConfig. This means we will not
be able to detect whether different ifmap's have different
configs for an object. Keeping bug open to address that later.

Change-Id: Ic25bc12b417b746d611d4f2064ae1c646f4d1571
Partial-Bug:1530299