Delete bgpaas-server object when VN is deleted
It is possible that one bgpaas-client object is being created while another one
is being deleted. To avoid race condition during this, we will leave the
bgpaas-server object when the client is deleted. These objects will be deleted
when the VN iteself is deleted. There are no resources held by such object, so
it should be ok to keep them even when they are not being used.

Also fixed a traceback in api server due to not catching cassandra
NotFoundException during update_last_modified.

Change-Id: Iaeea4b13d35443cd015b6b4e1145249d1bfc2606
Closes-Bug: 1529983