Create cassandra keyspaces from only one process
Use zookeeper to elect master for api server only for creating keyspaces in
cassandra. Also modified vnc_cassandra library so that the caller can specify
which keyspaces the caller owns so that other processes do not try to create
those keyspaces. Instead, those process will wait indefinitely for the shared
keyspaces to be created.

Change-Id: I29ce41a3a8025089cab973ad929f84aadc6d2abb
Closes-Bug: 1531023