Set expected_connections in query-engine based on collector conf
Handle collector connection in query-engine as below:
- If collector list is configured, then query-engine should connect to
one of the collectors in the list irrespective of whether discovery
service is configured or not.
- If collector list is not configured and discovery server is
configured, then query-engine should subscribe to collector service and
connect to one of the collectors provided by the discovery server.
- If both collector list and discovery server are not configured, then
connect to the local collector.

expected_connections in ConnectionStateManager::Init should be set
appropriately based on the collector and discovery server config.

Change-Id: Idee0fa9c3fba76ce106e2f3152b033af4b057d7d
Partial-Bug: #1528770