Fix corner case in routing instance deletion
Problem happens in the following scenario:

- Routing instance has service chain configuration
- Routing instance is on pending queue in service chain manager
- Routing instance is deleted but Shutdown method is not yet called
- Task trigger to resolve pending queue runs

At this time the routing instance has no configuration since it's been
marked deleted but it's still on the pending queue as Shutdown has not
been called. This makes ServiceChainMgr<T>::ResolvePendingServiceChain
dereference the routing instance's NULL config pointer.

This issue was introduced as part of code restructuring related to v6
service chaining. Fix by ensuring that the routing instance is removed
from the pending queue when it's marked deleted.

Credit to Prakash for figuring out the root cause.

Change-Id: I13617622a03360f1df30aa26396801adc704f50b
Closes-Bug: 1533438