Teach PathResolver to handle nexthop from arbitrary BgpTable
Generalize PathResolver so that it can track the nexthop address in a
BgpTable different than the one to which the BgpPath has been added.
Among other things, this would allow a inet6 path with an inet nexthop
to be resolved.

Highlights:

- Modify PathResolver APIs to take an optional BgpTable parameter
- ResolverNexthop keeps a delete reference to BgpTable in question
- ResolverNexthopMap is now keyed by (IpAddress, BgpTable)
- Handle corner case where BgpTable is marked deleted before the
ResolverNexthop can be registered to BgpConditionListener
- Add new test to path_resolver_test to exercise the above scenario
- Modify BgpTable::Input to use the new PathResolver APIs
- Maintain map of BgpTables keyed by address family in RoutingInstance
- Generalize path_resolver_test to test all 4 combinations of path
BgpTable and nexthop BgpTable for inet and inet6
- Add path_resolver_tests to bgp test suite

Change-Id: I3dc475cda5a79e0b6c13094ac733452a6600584d
Partial-Bug: 1503808