Send all fields of flow always in FlowLog messages
Currently vrouter-agent sends the following fields only once

setup-time
vrouter-ip
peer-vrouter-ip
underlay-port
tunnel-type

With flow sampling in place, if flow sample with above information is dropped,
the rest of the samples for this flow, which make it to collector will never
have this info. To overcome this we send all the fields always.

Also remove teardown-time since it is sent only once for delete and there are
chances of this sample getting dropped

Also remove unused flow fields from flow.sandesh.

Change-Id: Ib17fad30bc9d859375f82de363e57f97251686db
Closes-Bug: #1536118