Create vrouter state cache from polling uve
Currently every launch of a netns VM causes fetch of vrouter UVEs.
This is inefficient and the plan is to maintain a local cache for
the same. Also need to add debouncing logic for marking the vrouter
as down as the collectors might return inconsistent data when
there is flux in the system.

Change-Id: I81e77d9c975a0599d7036b19fd73e4042f3604a3
Closes-Bug: #1536821