Support IPv6 IP as mirror destination.
Create MirrorEntry pointing to MirrorNH even when vrf is empty. MirrorEntry in
this case was currently being made to point to DiscardNH. However if vrf string
is present and VrfObject is absent in agent then MirrorEntry will continue to
point to DiscardNH.

The tunnel NH of Vrouter can now take IPv6 IP as source and destination. When
IPv6 IP is being sent in tunnel NH set the family as AF_INET6. Replace
Ip4Address and inaddr types with IpAddress in Nexthop and Mirror Objects. The
source IP for MirrorNH when destination is IPv6 is V4 mapped V6 address of
vhost0.

Change-Id: If20293be84b5d94f547cf68d311d8e07d85657be
Closes-Bug: #1526427