Ignore change notification for loadbalancer-pool.
Ignore change notification for loadbalancer-pool when its ifmap-node is marked
for delete. During change notification handling for loadbalancer-pool we iterate
all its adjacent nodes. This iteration will result in crash if the entry is
marked for delete, because the graph vertex is reset as part of ifmap_node
deletion.

Also fix agent crash while converting fixed_ip to V4.

Change-Id: I6409e62e60952c71b7ced19d5a590e296b3b10aa
Closes-Bug: #1534348